Use intrusive holders for fallbacks in Chemical to ensure there are no performance hits in any of the places that have been moved to using holders

+++ src/api/java/mekanism/api/recipes/ingredients/creator/IChemicalStackIngredientCreator.java
@@ -23,7 +23,7 @@ public interface IChemicalStackIngredientCreator extends IIngredientCreator<Chem
     @SuppressWarnings("removal")
     @Deprecated(forRemoval = true, since = "10.7.11")
     default ChemicalStackIngredient from(Chemical instance, int amount) {
-        return fromHolder(instance.getAsHolder(), (long) amount);
+        return fromHolder(instance.getAsHolder(), amount);
     }
 
     @Override
